Make unit type implementation of BinRead clearer
The unit type impl as a special case of the recursive tuple macro
hid its existence within the macro itself. The doc-block tried to
make this clearer but described the invocation of the macro and
became out-of-sync. Moving the unit type impl out of the macro
entirely makes it clear that it is a special case.

See https://github.com/jam1garner/binread/pull/32 for more
discussion.