Improve top level attribute validation
There are currently too many ways that things are being validated
(parse_any validates the keyword is valid and can be parsed, then
Check validates the keyword is valid for the specific kind of
object being decorated, then a third time to check for duplicates
just before being transformed and assigned to the Attrs object).
So this probably needs to be simplified more.