Improve privacy and reduce redundancy in IdentStr code
The original purpose of most of this refactoring was actually to
cache the generated `TokenStream`s, since constantly regenerating
these takes a long time (an extra ~20000ns per call) and these
calls happen thousands of times.

However, it turns out that:

1. It not valid to retain any TokenStream between different
   invocations of a proc-macro. Trying will cause the compiler to
   crash (rust-lang/rust#66003);
2. `syn::Ident` has some problem where it will also cause the
   compiler to crash if it ends up in thread-local storage
   (rust-lang/rust#80473) and upstream seems to want to stop
   any thread-local storage in proc-macros eventually
   (rust-lang/rust#56058).

Being able to cache these tokens would provide a small but decent
speedup (~3%), so it is unfortunate that this is not possible.
Still, the formatting changes made while trying to implement a
cache seem reasonable enough to keep in the tree.