Disallow invalid field read mode directive combinations
This is a breaking change. In particular, this breaks the way that
`ignore` works, since despite the documentation saying it is an
alias for `default`, it actually would just do literally nothing
and therefore required an `import` with the same name as the field
to actually successfully build a struct. Since it is possible to
do that using `calc(var_name)`, that is now the correct blessed
way to pass values in through imports, and `ignore` now matches
what the documentation says it does and is just an alias for
`default`.