Restrict magic to types which actually implement BinRead
This also makes it possible to have a unit-like enum which use
byte arrays. It is trivially possible to change this to use a
standard enum instead, which adds some overhead (one read per
variant) but then all the extra features become available and only
`repr`-style enums then need a special code path, which might be
nicer.