Improve error output, especially for custom errors
This is a breaking change because it requires errors to implement
Display and Debug, but this is also the case for std::error::Error
so this should not be a problem for nearly every error type a user
will try to use.

It also requires magic to implement Debug, but magic must be a
primitive and all primitives implement Debug anyway so this part
is not a breaking change (though the output for byte arrays might
not be especially beautiful).

Refs jam1garner/binrw#9.