Call `args! {...}` from within the proc_macro
We discussed this on discord previously, it seems like the easiest
way to get this done without duplicating everything in the macro_rules
macro.

Vec<TokenStream> is meh, should probably adjust the type so it isn't
automatically a list, but this works for now.

This makes `args {...}` work for parse_with.