Use `Option<()>` instead of `bool` for keyword-only attributes.
Wrap keyword-only attribute types into a new `MetaVoid` type which
is convertible into `()`.

This makes the code more generic & consistent, no more specialized
implementation of `TrySet<bool>` & `TrySet<SpannedValue<bool>>`, no
more `Default` implementation for `SpannedValue<T>` which was only
used for `bool`. Having a default value for `SpannedValue` was a
bit inconsistent since the type is designed to wrap a span and in
the default case there is no span.