Fix improper copying of directives from enum to variant
That the original intent of copying directives from the enum to
each variant was probably to try to inherit things like assertions
and endianness, but the endianness is inherited normally just by
the way codegen works, and assertions from the enum are applied
explicitly to each variant, so it was not actually doing either of
these things. Instead, what it ended up doing was copying
directives that should *not* apply to each variant, like `magic`.
It also made it confusing which directives were actually necessary
for codegen to work correctly.

This patch changes the design so that variants are converted into a
struct-type input, which is simpler and matches more closely what
is actually happening.