Improve error reporting on args type mismatches
This commit provides an explicit message when someone combines
count + args in a way which is not valid, which was the primary
issue, and then also improves the diagnostic output for cases
where someone passes the wrong type of arguments to point to the
invalid args expression and the type of the field.

Notably, this does not resolve most issues with named arguments as
those are typically errors caused by calling non-existent builders
and builder functions, which requires a different strategy.

Closes jam1garner/binrw#112.