Prevent accidental use of native endianness
A fairly common problem with data serialisation occurs when code
that was intended to be cross-platform accidentally inherits the
endianness of the architecture used to develop the software.
This goes unnoticed until someone tries running it on a brand new
Motorola 68060 or some more modern CPU that still uses a big endian
architecture, like… uhhhhh…

Anyway, when network ASICs take over the world, they should not
have any accidental compatibility problems if they use binrw from
now on. This patch adds bounds to the endianless convenience
functions so they cannot be used except with types that have
explicitly defined endianness.

This patch also generalises the use of markers on binrw types so
that e.g. HasMagic is actually correctly specified for both read
and write implementations instead of only applying if someone
makes a read implementation. (The usefulness of this introspection
is still somewhat unclear to me though.)

One limitation of this patch is that it is impossible to correctly
derive the trait on objects where all the fields have the trait, so
someone e.g. wrapping a Vec<u8> would need to specify a useless
endianness directive or manually implement the
ReadEndian/WriteEndian traits.