Replace `Endian::Native` with cfg-defined `Endian::NATIVE` const
Native endianness is not actually a discrete kind of endianness,
so it does not really make sense as a variant. It also should not
be used except in very specific circumstances, like reading binary
data from a pipe or shared memory that is using a wire protocol
that does not explicitly define endianness.