Use a named feature for verbose backtraces instead of rustc version
If the verbose-backtrace feature is enabled (the default), then the
error display contains:

* Code at each frame location (currently only on nightly);
* Some pretty ANSI display codes;
* Some extra stuff in the emitted display output so you know for
  sure it’s a backtrace

Previously, some of these things (like having ANSI codes in the
errors and extra stuff in the output) were unconditionally emitted;
other things like showing code in backtrace frames were conditioned
exclusively on whether or not rustc was nightly.

This patch also enables code coverage over the backtrace code
which was previously disabled because I was too smoothbrain to
recognise that it only had to be gated on proc_macro.