Trigger error on multiple `try_optional` in NamedArgs
The codegen already does not handle this situation correctly
because it would require creating implementations for every
permutation of `try_optional` states but it only creates
implementations for each `try_optional` field. Since this attribute
cannot actually be used outside of binrw anyway due to compiler
restrictions, YAGNI this feature and just make sure no one tries to
do it anyway since it’ll be broken.