Mark `BufReader` as std-only in docs
doc_auto_cfg does not work correctly (all the methods of the
type’s impl inherit the marker but the struct itself does not) so
this patch just marks it manually.

Since doctests run with doc option set, it is necessary to also
gate on compiler that supports unstable features (nightly) since
this feature is not currently stabilised. This means binrw crate
gets a build.rs for a while. On the upside, jam1garner will
probably like how it eliminates a dev-dependency that is no longer
needed as a result. :-)