Revert `repr` on structs, enums, and fields
This feature does not work and fundamentally cannot work because
the write-side always receives references to values, but (1) there
are no `From<&T>` conversions for any std type that anyone would
use as a repr, and (2) this would cause a complete copy of the
converted object to be retained in memory just to write the object.
The `with` feature offers a more appropriate approach to custom
serialisations that does not suffer from these limitations.

This reverts:

"Implement top-level `repr` attribute for structs and enums."
commit 6f320039db1732014851e20285ecb6e4cbcb9d30

"Fix nightly tests"
commit 38b48d4f1fc2b996205ab2e0055f7401ac827db3

"Implement field-level `repr` for conversion."
commit 8353699f9640c1401802353dd97833608630585c

...with the exception of portions of those commits which fixed a
bug where arbitrary errors from `try_map` in BinWrite were not
allowed due to overly-restrictive type hints.