Improve semantics of `map_stream`
This change enables `map_stream` to be used in conjunction with
e.g. `count` on a field. This was not working before because the
fallible conversion code for `count` would try to access the
original stream to get the error position, and this was not
possible because the mapped stream had already been created and
was mutably borrowing the original stream.

After reflecting on the correct way to fix this issue, it also
seemed like it is counter-intuitive to have a mapped stream also
apply to sibling directives, so this patch changes the behaviour
so that the mapped stream is only used for reading the inner
value(s) of the thing it is applied to. This also reduces the scope
of the borrow and so should hopefully reduce the risk of any weird
interactions causing borrowck failures later.