Fix rustc inference issue with generic types in BinWrite
This issue was reported by a user in chat.

Using a generic type after a concrete type in a BinWrite
implementation causes the write function arguments type hint
to fail in a way that looks probably like a rustc bug. Instead of
correctly inferring `binrw::BinWrite::write_options` to be
`<T as binrw::BinWrite>::write_options`, it is inferred as if it
has the type of the previous field,
`<{concrete type} as binrw::BinWrite>::write_options`.

Passing the expected type explicitly to the hint function solves
the problem, though I think this means that auto-deref will not
work any more; the `if` directive had to be changed to not create
any double-reference, which did work previously.