Address edition 2024 UI test failures
The (necessary) use of function signature inference for the read
function, in the absence of an actually valid function to infer
from, results in the compiler generating warnings about `!` to `()`
decay behaviour changing in Rust 2024, even though this code is
already doomed to fail. We will see once edition 2024 is released
whether the compiler emits a diagnostic about `BinRead` not being
implemented for `!` or not.

(cherry picked from commit 4ca852152933e0b7ffda8b969280fff92a462e22)