Add SeekFrom::Start, SeekFrom::Current, and nested support to take_seek
This change adds support for the missing SeekFrom operations. It
refactors the internals a bit. We now keep a range of start..end values
that are accessible to the TakeSeek. I dropped manually computed stream
position and instead call the inner's `stream_position()`. This does
mean that `limit` needs to take in a `&mut`. I was also going to change
the return type to `Result<u64>`, but that would force all callers to
handle the `Result`, which I felt like it was a bigger API change.

I fixed the test added in 38b35989ffa27e8bb18cd91f69c6bcc66d2ca03e
because `StartFrom::Start(0)` used to reset the cursor to the start of
the underlying buffer.

Fixes jam1garner/binrw#291.