Update to syn version 2
As there were quite substantial changes from syn version 1 to 2 some
minor changes to the codebase are necessary:

The `Attribute` type is more accurately typed now, that means instead
of having simply a `Path` entity and some tokens after it, there is now
a differentation between attributes with a `=`, in parentheses or
without any other tokens but the path. This makes parsing slightly
easier for us. Some test fixes were also required as these rely on the
old tokens we got from an attribute.

The `Spanned` trait was sealed in syn v2. We implemented it for some
of the internal meta types. As we still have the span information
readily available and the trait is otherwise required nowhere, simply
remove the impl.

Some syntax enums were marked as `non-exhaustive`. Handle these
additional options in a sensible way where possible and throw compile
errors in all other places.