Ensure `TakeSeek::seek` uses correct stream end position
The purpose of `TakeSeek` is to be a binrw-compatible version of
`std::io::Take`, and the purpose of `std::io::Take` is only to
truncate the stream to a certain number of bytes, not to extend the
stream if a limit beyond the length of the inner stream is given.
Thus, `TakeSeek` also needs to pay attention to the true end of the
inner stream and use that as the `SeekFrom::End(0)` position if it
is less than the limit.

Refs jam1garner/binrw#291.