Update MSRV to 1.70
The trybuild dependency was fixed at an old version because one of
its newer dependencies updated MSRV to 1.70. Now, a second trybuild
dependency—which, ironically, was only used by trybuild for
compatibility with pre-1.70 Rust—has changed *its* MSRV to 1.70.

Even though binrw itself still does not require any 1.70+ features,
it is easier to update MSRV than it is to start committing and
updating Cargo.lock. If some users complain about the MSRV bump
then we can do the extra work and lower it again.