Allow combining `args` with `calc` for writes
Unlike `BinRead` where `calc` means the value is constructed with
no generated calls to `read_options`, for writing, `write_options`
is still called, so it is valid to pass arguments.

Thanks to @widberg for noticing this and reporting the issue and
for providing an excellent initial patch and minimal reproduction
case. By eliminating the extra code path for handling `calc` args
entirely, this patch subsumes the one in jam1garner/binrw#327.

Closes jam1garner/binrw#327.